+++ packages/tools/utils/CHANGELOG.md
@@ -8,6 +8,10 @@
 - renamed `reduceFromArr` -> `reduceFromEmptyArray`
 - renamed `reduceFromObj` -> `reduceFromEmptyObject`
 
+### (Any -> Object) -> (Array -> Object)
+
+- added makeAllOccurrencesWith
+
 ### Any -> Boolean
 
 - added `isArguments`